pyproject.toml
[project]
name = "tcc-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bleach>=6.2.0",
    "cryptography>=45.0.5",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "email-validator>=2.2.0",
    "psycopg2>=2.9.10",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "tzdata>=2025.2",
    "black",
    "isort",
    "ruff",
]

[tool.setuptools.packages.find]
include = ["accounts", "core", "domain", "restaurants", "users", "web"]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["accounts", "core", "domain", "restaurants", "users", "web"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D"]
extend-ignore = ["D"]


exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
